Rename mockValues to be test file
