+++ server/tests/controller/api/dashboardApi.test.ts
@@ -3,7 +3,7 @@ export {};
 import DashboardAPI from "../../../controller/api/dashboardApi";
 import DashboardParser from "../../../parser/dashboardParser";
 import { StatusCode } from "../../../types";
-import { FAKE_ERRORS, TEST_DASHBOARD } from "../global/mockValues";
+import { FAKE_ERRORS, TEST_DASHBOARD } from "../global/mockValues.test";
 jest.mock("../../../parser/dashboardParser");
 
 describe('Dashboard Functions', () => {
