+++ server/tests/controller/api/goalApi.test.ts
@@ -3,7 +3,7 @@ export {};
 import GoalAPI from "../../../controller/api/goalApi";
 import GoalParser from "../../../parser/goalParser";
 import { StatusCode } from "../../../types";
-import { FAKE_ERRORS, TEST_GOAL, TEST_SUB_GOAL } from "../global/mockValues";
+import { FAKE_ERRORS, TEST_GOAL, TEST_SUB_GOAL } from "../global/mockValues.test";
 import { GoalType } from "../../../types";
 jest.mock("../../../parser/goalParser");
