+++ server/tests/controller/api/loginApi.test.ts
@@ -4,7 +4,7 @@ import bcrypt from "bcryptjs";
 import LoginAPI from "../../../controller/api/loginApi";
 import LoginParser from "../../../parser/loginParser";
 import { StatusCode } from "../../../types";
-import { FAKE_ERRORS, TEST_ACCOUNT } from "../global/mockValues";
+import { FAKE_ERRORS, TEST_ACCOUNT } from "../global/mockValues.test";
 jest.mock("../../../parser/loginParser");
 
 describe('Login Functions', () => {
