+++ server/tests/controller/api/settingsApi.test.ts
@@ -3,7 +3,7 @@ export {};
 import SettingsApi from "../../../controller/api/settingsApi";
 import SettingsParser from "../../../parser/settingsParser";
 import { StatusCode } from "../../../types";
-import { FAKE_ERRORS, TEST_SETTINGS } from "../global/mockValues";
+import { FAKE_ERRORS, TEST_SETTINGS } from "../global/mockValues.test";
 jest.mock("../../../parser/settingsParser");
 
 describe('Settings Processor Unit Tests', () => {
