+++ server/tests/controller/processors/dashboardProcessor.test.ts
@@ -2,7 +2,7 @@ import * as DashboardProcessor from "../../../controller/processors/dashboardPro
 import DashboardAPI from "../../../controller/api/dashboardApi";
 import { StatusCode } from "../../../types";
 import { initializeErrorMap } from "../../../utils/errorMessages";
-import { createMockRequest, MOCK_RESPONSE, TEST_DASHBOARD } from "../global/mockValues";
+import { createMockRequest, MOCK_RESPONSE, TEST_DASHBOARD } from "../global/mockValues.test";
 
 jest.mock("../../../controller/api/dashboardApi");
