+++ server/tests/controller/processors/profileProcessor.test.ts
@@ -1,6 +1,6 @@
 import * as ProfileProcessor from "../../../controller/processors/profileProcessor";
 import ProfileAPI from "../../../controller/api/profileApi";
-import { createMockRequest, MOCK_RESPONSE, TEST_PROFILE, TEST_ACCOUNT } from "../global/mockValues";
+import { createMockRequest, MOCK_RESPONSE, TEST_PROFILE, TEST_ACCOUNT } from "../global/mockValues.test";
 import { StatusCode } from "../../../types";
 import { initializeErrorMap } from "../../../utils/errorMessages";
